refactor: (#798) 모범학생 조회 로직 정리 및 도메인 이동
