+++ dms-core/src/main/kotlin/team/aliens/dms/domain/student/service/GetStudentServiceImpl.kt
@@ -12,6 +12,9 @@ import team.aliens.dms.domain.student.exception.StudentNotFoundException
 import team.aliens.dms.domain.student.exception.StudentUpdateInfoNotFoundException
 import team.aliens.dms.domain.student.model.Student
 import team.aliens.dms.domain.student.spi.QueryStudentPort
+import team.aliens.dms.domain.vote.dto.response.ModelStudentListResponse
+import team.aliens.dms.domain.vote.exception.validateStudentList
+import java.time.LocalDate
 import java.util.UUID
 import java.util.function.Function
 
@@ -136,4 +139,26 @@ class GetStudentServiceImpl(
             }
         }
     }
+
+    override fun getModelStudentList(date: LocalDate): List<ModelStudentListResponse> {
+        val firstDayOfMonth = date.withDayOfMonth(1)
+        val lastDayOfMonth = date.withDayOfMonth(date.lengthOfMonth())
+
+        val startOfDay = firstDayOfMonth.atStartOfDay()
+        val endOfDay = lastDayOfMonth.atTime(23, 59, 59)
+
+        val modelStudentList = queryStudentPort.queryModelStudents(startOfDay, endOfDay)
+
+        validateStudentList(modelStudentList)
+
+        return modelStudentList.map { student ->
+            ModelStudentListResponse(
+                id = student.id,
+                gcn = student.studentGcn,
+                name = student.studentName,
+                profileImageUrl = student.studentProfile
+            )
+        }
+    }
+
 }
